refactor: modify auth middleware and adminController for test

+++ controllers/admin-controller.js
@@ -1,5 +1,6 @@
 const jwt = require("jsonwebtoken");
 const bcrypt = require("bcryptjs");
+const { getUser } = require("../helpers/auth-helper");
 const { User, Tweet, sequelize } = require("../models");
 const { QueryTypes } = require("sequelize");
 
@@ -38,7 +39,13 @@ const adminController = {
     }
   },
   getUsers: async (req, res, next) => {
+    const loginUser = getUser(req);
     try {
+      if (loginUser.role === "user") {
+        return res
+          .status(403)
+          .json({ status: "error", message: "permisson denied" });
+      }
       const users = await sequelize.query(
         `
       SELECT u.*, userTweet.tweetCount, userFollower.followerCount, userFollowing.followingCount, userTweetLike.userTweetLikeCount
@@ -86,7 +93,13 @@ const adminController = {
   },
   deleteTweet: async (req, res, next) => {
     const { id } = req.params;
+    const loginUser = getUser(req);
     try {
+      if (loginUser.role === "user") {
+        return res
+          .status(403)
+          .json({ status: "error", message: "permisson denied" });
+      }
       const deletedCount = await Tweet.destroy({ where: { id } });
       return res.json({ message: `刪除了 ${deletedCount} 筆資料` });
     } catch (error) {
